+++ tabletest-junit/src/test/java/io/github/nchaugen/tabletest/junit/converting/JavaConvertWithTest.java
@@ -2,10 +2,15 @@ package io.github.nchaugen.tabletest.junit.converting;
 
 import io.github.nchaugen.tabletest.junit.TableTest;
 import org.junit.jupiter.api.extension.ParameterContext;
+import org.junit.jupiter.params.converter.AnnotationBasedArgumentConverter;
 import org.junit.jupiter.params.converter.ArgumentConversionException;
 import org.junit.jupiter.params.converter.ArgumentConverter;
 import org.junit.jupiter.params.converter.ConvertWith;
 
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -18,6 +23,15 @@ public class JavaConvertWithTest {
         [name: Wilma, age: 19] | TEEN
         """)
     void testConvertWith(@ConvertWith(PersonConverter.class) Person person, AgeCategory expectedAgeCategory) {
+        assertEquals(expectedAgeCategory, person.ageCategory());
+    }
+
+    @TableTest("""
+        Person                 | AgeCategory?
+        [name: Fred, age: 22]  | ADULT
+        [name: Wilma, age: 19] | TEEN
+        """)
+    void testComposedConvertWith(@PersonType Person person, AgeCategory expectedAgeCategory) {
         assertEquals(expectedAgeCategory, person.ageCategory());
     }
 
@@ -35,21 +49,40 @@ public class JavaConvertWithTest {
             if (age < 20) return AgeCategory.TEEN;
             return AgeCategory.ADULT;
         }
+
+    }
+
+    private static Person convertToPerson(Object source) {
+        if (source instanceof Map attributes) {
+            return new Person(
+                (String) attributes.getOrDefault("name", "Fred"),
+                "Flintstone",
+                Integer.parseInt((String) attributes.getOrDefault("age", "16"))
+            );
+        }
+        throw new ArgumentConversionException("Cannot convert " + source.getClass().getSimpleName() + " to Person");
     }
 
     private static class PersonConverter implements ArgumentConverter {
-        @SuppressWarnings("rawtypes")
+        @SuppressWarnings({"rawtypes", "NullableProblems"})
         @Override
         public Object convert(Object source, ParameterContext context) throws ArgumentConversionException {
-            if (source instanceof Map attributes) {
-                return new Person(
-                    (String) attributes.getOrDefault("name", "Fred"),
-                    "Flintstone",
-                    Integer.parseInt((String) attributes.getOrDefault("age", "16"))
-                );
-            }
-            throw new ArgumentConversionException("Cannot convert " + source.getClass().getSimpleName() + " to Person");
+            return convertToPerson(source);
         }
+
     }
 
+    @Target({ElementType.ANNOTATION_TYPE, ElementType.PARAMETER, ElementType.FIELD})
+    @Retention(RetentionPolicy.RUNTIME)
+    @ConvertWith(PersonType.PersonConverter.class)
+    public @interface PersonType {
+
+        @SuppressWarnings("NullableProblems")
+        class PersonConverter extends AnnotationBasedArgumentConverter<PersonType> {
+            @Override
+            protected Object convert(Object source, Class<?> targetType, PersonType annotation) throws ArgumentConversionException {
+                return convertToPerson(source);
+            }
+        }
+    }
 }