+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Fixed
 - Published AsciiDoc files now uses `.adoc` extension instead of `.asciidoc`
+- Detects `@ConvertWith` parameter annotation when used in custom composed annotations
 
 ## [0.5.4] - 2025-11-03
 ### Added